requirements.txt
# Core environment
openenv>=0.2.3
gymnasium==1.3.0
networkx==3.4.2
numpy==2.2.4
pydantic==2.12.5
pydantic-settings==2.8.1
PyYAML==6.0.2

# API server
fastapi==0.115.12
uvicorn==0.42.0
starlette==0.46.1
httpx==0.28.1
anyio==4.9.0

# Dashboard
gradio==6.10.0
gradio_client==2.4.0
pandas==2.2.3

# ML / Training (optional — comment out if no GPU)
torch==2.6.0
transformers==4.51.1
# trl and unsloth are intentionally optional due to GPU/platform constraints.
# Install for full GRPO training runs:
# pip install trl unsloth

# Vector DB
# chromadb>=0.4.0  # install separately: pip install chromadb

# Testing
pytest==9.0.2
hypothesis==6.152.2

# Utilities
requests==2.32.3

# LLM APIs — HuggingFace Inference + OpenAI Oracle
huggingface_hub>=0.22.0
openai>=1.0.0
python-dotenv>=1.0.0
